% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gstest_graph.R
\name{gen_gs_graph}
\alias{gen_gs_graph}
\title{gen_gs_graph
Create adjacency matrix for a graph with edge probability and}
\usage{
gen_gs_graph(n, pr_edge, nloops, loop_size)
}
\description{
gen_gs_graph
Create adjacency matrix for a graph with edge probability and
}
