% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gs_result.R
\name{get_gs_result}
\alias{get_gs_result}
\title{get_gs_result}
\usage{
get_gs_result(file_name, node_names = NULL)
}
\arguments{
\item{file_name}{prefix for file names containing edge probabilities (\code{_edge_p.out} appended to \code{file_name}),
best graph (\code{_best_graph.out}), graph samples (\code{_graph_samples.out})}
}
\value{
A list object of class \code{rgs} containing edge probabilities, best graph, samples, SCC probabilities.
}
\description{
Load information from graph_sampler output files into a single R object
}
\details{
Loading of samples and SCCs can also be accomplished on its own using
\code{get_samples} and \code{get_scc} functions.
This function is a "wrapper" for loading all relevant data.
}
\author{
Witold Wiecek
}
