% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{get_samples
Turn a graph_sampler samples output file into a list in R.
See \code{get_gs_result} for a more general function for working with outputs.}
\usage{
get_samples(filename = "graph_samples.out")
}
\arguments{
\item{filename}{graph samples (adjacency matrices) file from \emph{graph_sampler}
(typically file names ending with \code{graph_samples.out})}
}
\value{
List of adjacency matrices
}
\description{
get_samples
Turn a graph_sampler samples output file into a list in R.
See \code{get_gs_result} for a more general function for working with outputs.
}
