% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scc.R
\name{get_scc}
\alias{get_scc}
\title{get_scc
Turn a \code{graph_sampler} samples output file into a list of strongly connected components.
See \code{get_gs_result} for a more general function for working with outputs.}
\usage{
get_scc(filename = "graph_samples.out")
}
\arguments{
\item{filename}{graph samples (adjacency matrices) file from \emph{graph_sampler}
(typically file names ending with \code{graph_samples.out})}
}
\value{
List of strongly connected components
}
\description{
get_scc
Turn a \code{graph_sampler} samples output file into a list of strongly connected components.
See \code{get_gs_result} for a more general function for working with outputs.
}
