% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rgs.R
\name{print.rgs}
\alias{print.rgs}
\title{Generic print for rgs objects}
\usage{
\method{print}{rgs}(x, ...)
}
\arguments{
\item{x}{Object of class \code{rgs}, typically produced by \code{\link[=rgs]{rgs()}} or \code{\link[=get_gs_result]{get_gs_result()}}.}
}
\description{
Generic print for rgs objects
}
