% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgraphsampler.R
\docType{package}
\name{rgraphsampler-package}
\alias{rgraphsampler}
\alias{rgraphsampler-package}
\title{\code{graph_sampler} is a network inference and simulation package developed by Frederic Bois.
\code{graph_sampler} allows you to generate structured random graphs and perform fully Bayesian
inference on Bayesian networks' or dynamic Bayesian networks' structure. If you are not
familiar with \code{graph_sampler}, please refer to the
\href{http://www.nongnu.org/graphsampler/}{project's website}.}
\description{
\code{rgraphsampler} is a simple R interface for calling \code{graph_sampler}'s C code (v3.1) from R
and manipulating the results, designed primarily for non-Linux users.
}
\details{
The objective of \code{rgraphsampler} is two-fold:
\enumerate{
\item To make available to R users a convenient tool for performing inference in \code{graph_sampler}
\item To make working with outputs of \code{graph_sampler} analyses easier.
}
}
\section{Basic use case}{


\code{graph_sampler} scripts are ran with \code{rgs()}; \code{rgs(file_in, file_out)} is equivalent
to \code{graph_sampler <file_in> <file_out>}
See \code{?rgs} for examples and \code{vignette("rgraphsampler")} for a short demonstration.
}

\section{Differences between R and C versions}{


R version is fully featured in that it directly uses unmodified \code{graph_sampler} code within
the package. However, GSL versions of functions are disabled to decrease dependencies for non-Linux users.
Therefore the results obtained through R version should be equivalent to running non-GSL version
of \code{graph_sampler}.
}

\section{Known bugs, state of the project}{


Due to nature of \code{bison} parsing engine (the tool for reading in input scripts), when errors
occur (e.g. due to malformed input script) in the parser, it is recommended to unload the
dynamic library before working with any other scripts. This helps R delete some static variables
generated by the parser, which otherwise may not reset on errors. To unload the namespace use
\code{unloadNamespace("rgraphsampler")}

All of the code included here has been tested with the scripts from \code{graph_sampler}'s (C version)
testing suite. That means this package produces outputs identical to GSL version of \code{graph_sampler}
that we've tested. However, there is no automated testing of R code and caution is advised.

#' @keywords internal
}

\author{
\strong{Maintainer}: Witold Wiecek \email{witold.wiecek@gmail.com}

Authors:
\itemize{
  \item Frederic Bois
}

}
