% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgs.R
\name{rgs}
\alias{rgs}
\title{rgs: R graph_sampler interface}
\usage{
rgs(file_in = "script.txt", file_out = "output")
}
\arguments{
\item{file_in}{file name for a valid graphsampler input script}

\item{file_out}{file name (prefix) for a valid}
}
\value{
Nothing is returned; standard graph_sampler output is printed
        in the terminal and output files generated as requested.
}
\description{
rgs: R graph_sampler interface
}
\details{
This simple function allows running C software graph_sampler
         inside R. All of the graph_sampler configuration is achieved
         by using a config script (`file_in`), which is identical to
         9the input for C version)[http://www.nongnu.org/graphsampler/graph_sampler.html#Running-Graph_005fSampler]
}
\author{
Witold Wiecek
}
