% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{summary.rgs}
\alias{summary.rgs}
\title{Generic summary for rgs objects}
\usage{
\method{summary}{rgs}(object, ..., digits)
}
\arguments{
\item{object}{Object of class \code{rgs}, typically produced by \code{\link[=rgs]{rgs()}} or \code{\link[=get_gs_result]{get_gs_result()}}.}
}
\description{
Generic summary for rgs objects
}
