% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgs.R
\name{rgs}
\alias{rgs}
\title{rgs: R graph_sampler interface}
\usage{
rgs(file_in = "script.txt", file_out = "output")
}
\arguments{
\item{file_in}{file name for a valid \code{graph_sampler} input script}

\item{file_out}{file name (prefix) for output files generated by \code{graph_sampler}}
}
\value{
Nothing is returned; standard graph_sampler output is printed
in the terminal and output files generated as requested.
}
\description{
Run \href{http://www.nongnu.org/graphsampler/}{graph_sampler} software from inside R.
}
\details{
This simple function allows running C software graph_sampler
inside R. All of the graph_sampler configuration is achieved
by using a config script (\code{file_in}), which is identical to
\href{http://www.nongnu.org/graphsampler/graph_sampler.html#Running-Graph_005fSampler}{the input for C version}
}
\examples{
script_file <- system.file(package="rgraphsampler",
                           "test_scripts", "script_test_1.txt")
rgs(file_in = script_file,
    file_out = "gs_output")

}
\author{
Witold Wiecek
}
