% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgs.R
\name{rgs}
\alias{rgs}
\title{rgs: R graph_sampler interface}
\usage{
rgs(file_in = "script.txt", file_out = "output",
  return_result = TRUE, node_names = NULL)
}
\arguments{
\item{file_in}{file name for a valid \code{graph_sampler} input script (see syntax
\href{http://www.nongnu.org/graphsampler/graph_sampler.html#Input-file-syntax}{here};
comprehensive list of examples is also attached -- see Details)}

\item{file_out}{file name (prefix) for output files generated by \code{graph_sampler}}

\item{return_result}{logical; if TRUE, call to \code{get_gs_result()} is made at the end of the function}

\item{node_names}{passed to \code{get_gs_result()} if \code{return_result = TRUE}}
}
\value{
If \code{return_result = FALSE}, nothing is returned; standard graph_sampler output is printed
in the terminal and output files generated as requested, under \code{file_out} file name.
If \code{return_result = TRUE}, \code{\link[=get_gs_result]{get_gs_result()}} is ran after \code{graph_sampler} and the
output is an \code{rgs} object.
}
\description{
Run \href{http://www.nongnu.org/graphsampler/}{graph_sampler} software from inside R.
\code{rgs(file_in, file_out)} is equivalent to \code{graph_sampler <file_in> <file_out>}
}
\details{
This simple function allows running C software graph_sampler
inside R. All of the graph_sampler configuration is achieved
by using a config script (\code{file_in}), which is identical to
\href{http://www.nongnu.org/graphsampler/graph_sampler.html#Running-Graph_005fSampler}{the input for C version}.
30 examples of scripts are attached within this package.
They can be accessed by \code{system.file()}, like in the example below.
For the full list use \code{list.files(system.file(package="rgraphsampler", "test_scripts"))}.
}
\examples{
script_file <- system.file(package="rgraphsampler",
                           "test_scripts", "script_test_1.txt")
# to review built-in examples try
# file.show(script_file)

# Will generate graph_sampler output files in the current working directory
rgs(file_in = script_file,
    file_out = "gs_output")

}
\author{
Witold Wiecek
}
